Adding QuickNav Customization
Adding a way to customize the quick navigation buttons in the inventory.
Every button can be toggled, has a display icon with nbt data, a click event and GUI name for highlighting.

On default, everything is like before.

The only problem is cloth config translation. I never worked before with cloth and don't want to copy paste the same key 12 times for the 12 buttons.
If anyone is good with cloth config and know how to change the path, just edit it or message me :+1: