fixes #539 #538
The logic of the livid code has been changed.
The system checks whether the current color is the same as the last found livid color. The initial color is detected after the first blindness effect + 2 seconds.
I also added the bounding box option for livid to make it easier to recognize the correct enemy. I changed default mob glow option to false by and bounding box option to true.