Fix regex and change sent message format
This should allow skyhanni users to be able to parse the messages sent by skyblocker users since their regex starts matching from the start of the message content (so after any prefixes, player names)

The regex change was also because of this. We use .find so the regex doesn't have to match the whole message body.