Add customizable night vision strength feature (#1078)
* Add customizable night vision strength feature

Introduced a `nightVision` subcommand to adjust night vision strength with options for `full` (= 100), `off` (= 0), and any other value between 0 and 100. Integrated the setting into the configuration menu and game rendering, with default strength set to 100. Updated localization to include feedback for strength changes.

* Update GameRendererMixin.java

Clamped the applied value to ensure it's not overflowing.
Although you can't put an invalid value in-game, it's still possible to do so directly via the config file.

* Update en_us.json

Added the missing translation for the night vision strength option.

* Add description tooltip for night vision strength option

A tooltip was added to the "Night Vision Strength" option to improve user understanding of its functionality. Unused import(s) were also removed, following the requested changes.

* Use clamp

* Apply suggestions from code review

---------

Co-authored-by: Manchick0 <endercreep109@gmail.com>
Co-authored-by: Kevin <92656833+kevinthegreat1@users.noreply.github.com>