Add Speed Presets for configurable speed settings (#1111)
* Add Speed Presets for configurable speed settings

Introduced a Speed Presets system with customizable presets for the rancher boots. Added a GUI to manage presets and integrated support for the `setmaxspeed` command, and the sign-editor to use these presets.

Besides all of that, refactored item protection rendering logic to show it everywhere, where applicable, and render it in front of the item, as well as updated the texture being used.

* Reverted the proposed changes to the protected item texture

I apologize for the changes, since I didn't know that the texture was applied elsewhere. This commit reverts all the changes to the size of the texture and the way it is rendered, to hopefully prevent any issues.

* Apply suggestions from code review

Co-authored-by: Kevin <92656833+kevinthegreat1@users.noreply.github.com>

* Refactor modification checks for SpeedPreset entries.

Removed the `hasBeenModified` flag, and replaced it with direct comparisons of initial and current values.

* Clean up pointless changes

* Refactor SpeedPresets to use Object2IntMap.

Replaced LinkedHashMap with Object2IntMap for efficiency and consistency in managing speed presets. Adjusted related methods and logic to accommodate the new data structure. Updated SpeedPresetListWidget to handle changes and comparisons using the revised map implementation.

* Refactor SpeedPresets

---------

Co-authored-by: Kevin <92656833+kevinthegreat1@users.noreply.github.com>