ItemList changes
This update introduces a bunch of changes to the ItemList feature, to hopefully enhance user experience. Most importantly, you can now filter out specific elements in the list to narrow down the results. The lore-search now works properly, and you're now able to use the wiki lookup key in the recipe book. Besides all that, shift-clicking on the recipe result button will suggest the recipe to the player using the `viewrecipe` command, to allow quick super-craft access.