ItemList Changes (#1119)
* ItemList changes

This update introduces a bunch of changes to the ItemList feature, to hopefully enhance user experience. Most importantly, you can now filter out specific elements in the list to narrow down the results. The lore-search now works properly, and you're now able to use the wiki lookup key in the recipe book. Besides all that, shift-clicking on the recipe result button will suggest the recipe to the player using the `viewrecipe` command, to allow quick super-craft access.

* Add missing GUI filter icons

* Update FilterOption.java

Properly use parentheses in the `(animal)` and `(pest)` flagged entries.

* Refactor recipe view command to use MessageScheduler.

Replaces direct chat command handling with MessageScheduler.

* Refactor filtering logic and remove Identifiable interface.

Replaced the Identifiable interface with a Supplier-based design for enum, updated related classes to use Supplier methods and refactored filtering logic for consistency. Additionally, added a new tip for viewing recipes in the recipe book, and changed the Shift+Click requirement to a simple right click.

* Clean up the code