Use custom vanilla GSON config serialization
The mysterious config wipe errors were the result of malformed json,
normal GSON is extensively tested and is highly unlikely to produce that
itself but YACL uses an intermediary parser that has had similar
problems with this stuff in the past.

By circumventing that library, this issue can hopefully be fixed.