Add more shortcuts
I use a QWERTY keyboard. I try to choose the letters for the abbreviations based on the characters I would subconsciously type when typing the full path.

Some shortcuts that don't follow this rule are usually because there are duplicate parameters or they can be entered quickly by double-clicking.

Most warp commands start with "w" to limit them to the same level, so that the auto-complete function does not provide too much warp commands.

By listing these commands, even if the user does not like the shortcuts, they can quickly change it without having to look up the specific command to enter.