Add MP to PV
Closes #1924

Added magical power calculation and display to profile viewer UI.

**Changes:**
- added 'getMagicalPower(JsonObject playerProfile)' function to calculate MP from accessories
- display MP in '/pv' ('SkillsPage') UI
- profile viewer UI increased height by 8 pixels

**Testing:**
Tested it with random Skyblock profiles from early game to maxed.
For comparison I used the Skycrypt MP calculation / data.

**Problems:**
The algorithm can be off for a few points at certain profiles because 'ItemStack.getSkyblockRarity()' returns 'SPECIAL' for 'VERY_SPECIAL' Hatcessories.
So it can happen, that the player has 1817 but Skyblocker shows 1815 MP because it sees 'VERY_SPECIAL' as 'SPECIAL'.
Max difference because of this is 4 MP, because you can only have 2 active 'VERY_SPECIAL' Hatcessories.